Move callbacks into PipelineRunContext
