+++ graphrag/index/operations/summarize_descriptions/description_summary_extractor.py
@@ -6,7 +6,7 @@
 import json
 from dataclasses import dataclass
 
-from graphrag.index.typing import ErrorHandlerFn
+from graphrag.index.typing.error_handler import ErrorHandlerFn
 from graphrag.index.utils.tokens import num_tokens_from_string
 from graphrag.language_model.protocol.base import ChatModel
 from graphrag.prompts.index.summarize_descriptions import SUMMARIZE_PROMPT
