+++ graphrag/index/run/utils.py
@@ -5,10 +5,13 @@
 
 from graphrag.cache.memory_pipeline_cache import InMemoryCache
 from graphrag.cache.pipeline_cache import PipelineCache
+from graphrag.callbacks.noop_workflow_callbacks import NoopWorkflowCallbacks
 from graphrag.callbacks.progress_workflow_callbacks import ProgressWorkflowCallbacks
 from graphrag.callbacks.workflow_callbacks import WorkflowCallbacks
 from graphrag.callbacks.workflow_callbacks_manager import WorkflowCallbacksManager
-from graphrag.index.context import PipelineRunContext, PipelineRunStats, PipelineState
+from graphrag.index.typing.context import PipelineRunContext
+from graphrag.index.typing.state import PipelineState
+from graphrag.index.typing.stats import PipelineRunStats
 from graphrag.logger.base import ProgressLogger
 from graphrag.storage.memory_pipeline_storage import MemoryPipelineStorage
 from graphrag.storage.pipeline_storage import PipelineStorage
@@ -17,6 +20,7 @@ from graphrag.storage.pipeline_storage import PipelineStorage
 def create_run_context(
     storage: PipelineStorage | None = None,
     cache: PipelineCache | None = None,
+    callbacks: WorkflowCallbacks | None = None,
     stats: PipelineRunStats | None = None,
     state: PipelineState | None = None,
 ) -> PipelineRunContext:
@@ -25,6 +29,7 @@ def create_run_context(
         stats=stats or PipelineRunStats(),
         cache=cache or InMemoryCache(),
         storage=storage or MemoryPipelineStorage(),
+        callbacks=callbacks or NoopWorkflowCallbacks(),
         state=state or {},
     )
