+++ graphrag/index/typing/context.py
@@ -2,37 +2,18 @@
 # Licensed under the MIT License
 
 # isort: skip_file
-"""A module containing the 'PipelineRunStats' and 'PipelineRunContext' models."""
+"""A module containing the 'PipelineRunContext' models."""
 
-from dataclasses import dataclass as dc_dataclass
-from dataclasses import field
-from typing import Any
+from dataclasses import dataclass
 
 from graphrag.cache.pipeline_cache import PipelineCache
+from graphrag.callbacks.workflow_callbacks import WorkflowCallbacks
+from graphrag.index.typing.state import PipelineState
+from graphrag.index.typing.stats import PipelineRunStats
 from graphrag.storage.pipeline_storage import PipelineStorage
 
 
-@dc_dataclass
-class PipelineRunStats:
-    """Pipeline running stats."""
-
-    total_runtime: float = field(default=0)
-    """Float representing the total runtime."""
-
-    num_documents: int = field(default=0)
-    """Number of documents."""
-
-    input_load_time: float = field(default=0)
-    """Float representing the input load time."""
-
-    workflows: dict[str, dict[str, float]] = field(default_factory=dict)
-    """A dictionary of workflows."""
-
-
-PipelineState = dict[Any, Any]
-
-
-@dc_dataclass
+@dataclass
 class PipelineRunContext:
     """Provides the context for the current pipeline run."""
 
@@ -41,5 +22,7 @@ class PipelineRunContext:
     "Long-term storage for pipeline verbs to use. Items written here will be written to the storage provider."
     cache: PipelineCache
     "Cache instance for reading previous LLM responses."
+    callbacks: WorkflowCallbacks
+    "Callbacks to be called during the pipeline run."
     state: PipelineState
     "Arbitrary property bag for runtime state, persistent pre-computes, or experimental features."
