+++ graphrag/index/typing/state.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2024 Microsoft Corporation.
+# Licensed under the MIT License
+
+"""Pipeline state types."""
+
+from typing import Any
+
+PipelineState = dict[Any, Any]
