+++ graphrag/index/workflows/create_final_text_units.py
@@ -5,11 +5,10 @@
 
 import pandas as pd
 
-from graphrag.callbacks.workflow_callbacks import WorkflowCallbacks
 from graphrag.config.models.graph_rag_config import GraphRagConfig
 from graphrag.data_model.schemas import TEXT_UNITS_FINAL_COLUMNS
-from graphrag.index.context import PipelineRunContext
-from graphrag.index.typing import WorkflowFunctionOutput
+from graphrag.index.typing.context import PipelineRunContext
+from graphrag.index.typing.workflow import WorkflowFunctionOutput
 from graphrag.utils.storage import (
     load_table_from_storage,
     storage_has_table,
@@ -20,7 +19,6 @@ from graphrag.utils.storage import (
 async def run_workflow(
     config: GraphRagConfig,
     context: PipelineRunContext,
-    _callbacks: WorkflowCallbacks,
 ) -> WorkflowFunctionOutput:
     """All the steps to transform the text units."""
     text_units = await load_table_from_storage("text_units", context.storage)
