+++ graphrag/index/workflows/finalize_graph.py
@@ -8,19 +8,18 @@ import pandas as pd
 from graphrag.callbacks.workflow_callbacks import WorkflowCallbacks
 from graphrag.config.models.embed_graph_config import EmbedGraphConfig
 from graphrag.config.models.graph_rag_config import GraphRagConfig
-from graphrag.index.context import PipelineRunContext
 from graphrag.index.operations.create_graph import create_graph
 from graphrag.index.operations.finalize_entities import finalize_entities
 from graphrag.index.operations.finalize_relationships import finalize_relationships
 from graphrag.index.operations.snapshot_graphml import snapshot_graphml
-from graphrag.index.typing import WorkflowFunctionOutput
+from graphrag.index.typing.context import PipelineRunContext
+from graphrag.index.typing.workflow import WorkflowFunctionOutput
 from graphrag.utils.storage import load_table_from_storage, write_table_to_storage
 
 
 async def run_workflow(
     config: GraphRagConfig,
     context: PipelineRunContext,
-    callbacks: WorkflowCallbacks,
 ) -> WorkflowFunctionOutput:
     """All the steps to create the base entity graph."""
     entities = await load_table_from_storage("entities", context.storage)
@@ -29,7 +28,7 @@ async def run_workflow(
     final_entities, final_relationships = finalize_graph(
         entities,
         relationships,
-        callbacks,
+        callbacks=context.callbacks,
         embed_config=config.embed_graph,
         layout_enabled=config.umap.enabled,
     )
