+++ tests/verbs/test_create_base_text_units.py
@@ -1,7 +1,6 @@
 # Copyright (c) 2024 Microsoft Corporation.
 # Licensed under the MIT License
 
-from graphrag.callbacks.noop_workflow_callbacks import NoopWorkflowCallbacks
 from graphrag.config.create_graphrag_config import create_graphrag_config
 from graphrag.index.workflows.create_base_text_units import run_workflow
 from graphrag.utils.storage import load_table_from_storage
@@ -22,11 +21,7 @@ async def test_create_base_text_units():
 
     config = create_graphrag_config({"models": DEFAULT_MODEL_CONFIG})
 
-    await run_workflow(
-        config,
-        context,
-        NoopWorkflowCallbacks(),
-    )
+    await run_workflow(config, context)
 
     actual = await load_table_from_storage("text_units", context.storage)
 
@@ -46,11 +41,7 @@ async def test_create_base_text_units_metadata():
 
     await update_document_metadata(config.input.metadata, context)
 
-    await run_workflow(
-        config,
-        context,
-        NoopWorkflowCallbacks(),
-    )
+    await run_workflow(config, context)
 
     actual = await load_table_from_storage("text_units", context.storage)
     compare_outputs(actual, expected)
@@ -70,11 +61,7 @@ async def test_create_base_text_units_metadata_included_in_chunk():
 
     await update_document_metadata(config.input.metadata, context)
 
-    await run_workflow(
-        config,
-        context,
-        NoopWorkflowCallbacks(),
-    )
+    await run_workflow(config, context)
 
     actual = await load_table_from_storage("text_units", context.storage)
     # only check the columns from the base workflow - our expected table is the final and will have more
