+++ tests/verbs/test_create_communities.py
@@ -1,7 +1,6 @@
 # Copyright (c) 2024 Microsoft Corporation.
 # Licensed under the MIT License
 
-from graphrag.callbacks.noop_workflow_callbacks import NoopWorkflowCallbacks
 from graphrag.config.create_graphrag_config import create_graphrag_config
 from graphrag.data_model.schemas import COMMUNITIES_FINAL_COLUMNS
 from graphrag.index.workflows.create_communities import (
@@ -32,7 +31,6 @@ async def test_create_communities():
     await run_workflow(
         config,
         context,
-        NoopWorkflowCallbacks(),
     )
 
     actual = await load_table_from_storage("communities", context.storage)
