+++ tests/verbs/test_create_final_documents.py
@@ -1,7 +1,6 @@
 # Copyright (c) 2024 Microsoft Corporation.
 # Licensed under the MIT License
 
-from graphrag.callbacks.noop_workflow_callbacks import NoopWorkflowCallbacks
 from graphrag.config.create_graphrag_config import create_graphrag_config
 from graphrag.data_model.schemas import DOCUMENTS_FINAL_COLUMNS
 from graphrag.index.workflows.create_final_documents import (
@@ -27,11 +26,7 @@ async def test_create_final_documents():
 
     config = create_graphrag_config({"models": DEFAULT_MODEL_CONFIG})
 
-    await run_workflow(
-        config,
-        context,
-        NoopWorkflowCallbacks(),
-    )
+    await run_workflow(config, context)
 
     actual = await load_table_from_storage("documents", context.storage)
 
@@ -54,11 +49,7 @@ async def test_create_final_documents_with_metadata_column():
 
     expected = await load_table_from_storage("documents", context.storage)
 
-    await run_workflow(
-        config,
-        context,
-        NoopWorkflowCallbacks(),
-    )
+    await run_workflow(config, context)
 
     actual = await load_table_from_storage("documents", context.storage)
