+++ tests/verbs/test_pipeline_state.py
@@ -3,26 +3,24 @@
 
 """Tests for pipeline state passthrough."""
 
-from graphrag.callbacks.noop_workflow_callbacks import NoopWorkflowCallbacks
-from graphrag.callbacks.workflow_callbacks import WorkflowCallbacks
 from graphrag.config.create_graphrag_config import create_graphrag_config
 from graphrag.config.models.graph_rag_config import GraphRagConfig
-from graphrag.index.context import PipelineRunContext
 from graphrag.index.run.utils import create_run_context
-from graphrag.index.typing import WorkflowFunctionOutput
+from graphrag.index.typing.context import PipelineRunContext
+from graphrag.index.typing.workflow import WorkflowFunctionOutput
 from graphrag.index.workflows.factory import PipelineFactory
 from tests.verbs.util import DEFAULT_MODEL_CONFIG
 
 
 async def run_workflow_1(  # noqa: RUF029
-    _config: GraphRagConfig, context: PipelineRunContext, _callbacks: WorkflowCallbacks
+    _config: GraphRagConfig, context: PipelineRunContext
 ):
     context.state["count"] = 1
     return WorkflowFunctionOutput(result=None)
 
 
 async def run_workflow_2(  # noqa: RUF029
-    _config: GraphRagConfig, context: PipelineRunContext, _callbacks: WorkflowCallbacks
+    _config: GraphRagConfig, context: PipelineRunContext
 ):
     context.state["count"] += 1
     return WorkflowFunctionOutput(result=None)
@@ -35,11 +33,10 @@ async def test_pipeline_state():
 
     config = create_graphrag_config({"models": DEFAULT_MODEL_CONFIG})
     config.workflows = ["workflow_1", "workflow_2"]
-    callbacks = NoopWorkflowCallbacks()
     context = create_run_context()
 
     for _, fn in PipelineFactory.create_pipeline(config).run():
-        await fn(config, context, callbacks)
+        await fn(config, context)
 
     assert context.state["count"] == 2
 
@@ -49,10 +46,9 @@ async def test_pipeline_existing_state():
 
     config = create_graphrag_config({"models": DEFAULT_MODEL_CONFIG})
     config.workflows = ["workflow_2"]
-    callbacks = NoopWorkflowCallbacks()
     context = create_run_context(state={"count": 4})
 
     for _, fn in PipelineFactory.create_pipeline(config).run():
-        await fn(config, context, callbacks)
+        await fn(config, context)
 
     assert context.state["count"] == 5
