+++ tests/verbs/util.py
@@ -5,8 +5,8 @@ import pandas as pd
 from pandas.testing import assert_series_equal
 
 import graphrag.config.defaults as defs
-from graphrag.index.context import PipelineRunContext
 from graphrag.index.run.utils import create_run_context
+from graphrag.index.typing.context import PipelineRunContext
 from graphrag.utils.storage import load_table_from_storage, write_table_to_storage
 
 pd.set_option("display.max_columns", None)
